+++ src/features/auth/index.ts
@@ -1,5 +1,4 @@
 export {default as useLoginModal} from './lib/useLoginModal';
-export {default as useKakaoOAuthLogin} from './lib/useKakaoOAuthLogin';
 export {default as useGoLoginPage} from './lib/useGoLoginPage';
 export {default as LoginButton} from './ui/LoginButton';
 export {default as LoginButtonLoading} from './ui/LoginButtonLoading';
